requirements.txt
discord.py==1.7.3
nekos.py==1.0.3